refactor(map): reorganize OpenGL classes into a structured package
Moved OpenGL-related classes (e.g., `Star`, `MyGLRenderer`, `Camera`) to the `model.map` package for improved modularity.
Updated `Map.kt` to use the new package structure.
Removed commented code in MyGLRenderer.kt
