+++ app/src/main/java/com/github/lookupgroup27/lookup/model/map/MyGLSurfaceView.kt
@@ -1,4 +1,4 @@
-package com.github.lookupgroup27.lookup.opengl
+package com.github.lookupgroup27.lookup.model.map
 
 import android.content.Context
 import android.opengl.GLSurfaceView
